+++ sdl3-sys/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changes
 
+- next:
+    - Add `from_raw` assoc function and `into_raw` and `value` methods for groups (ids, enums, etc),
+      and impl Display for ids under the `display-impls` feature
+
 - 0.6.0:
     - Update SDL to 3.4.0
     - Make IDs and flag types newtypes