+++ sdl3-ttf-sys/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changes
 
+- next:
+    - Add `from_raw` assoc function and `into_raw` and `value` methods for groups (ids, enums, etc),
+      and impl Display for ids under the `display-impls` feature
+
 - 0.6.0:
     - Update sdl3-sys to 0.6.0
     - Synchronize version number with sdl3-sys